feat: add ENVIRONMENT to workers; implement worker health check in App.vue
