+++ src/stores/blockchain/chains.ts
@@ -4,18 +4,18 @@ export type CHAIN_ID = MAINNET_CHAIN_ID | TESTNET_CHAIN_ID
 
 export enum MAINNET_CHAIN_ID {
 	ETHEREUM = '1',
-	POLYGON = '137',
-	OPTIMISM = '10',
-	ARBITRUM = '42161',
 	BASE = '8453',
+	ARBITRUM = '42161',
+	OPTIMISM = '10',
+	POLYGON = '137',
 }
 
 export enum TESTNET_CHAIN_ID {
 	LOCAL = '1337',
 	SEPOLIA = '11155111',
+	BASE_SEPOLIA = '84532',
 	ARBITRUM_SEPOLIA = '421614',
 	OPTIMISM_SEPOLIA = '11155420',
-	BASE_SEPOLIA = '84532',
 	POLYGON_AMOY = '80002',
 }
